hw4_parallel_programming/requirements.txt
pillow==9.0.1
hurry.filisize==0.9
